update to put functions to accept larger datatypes

+++ incl/put.h
@@ -6,7 +6,7 @@
 /*   By: rlutt <rlutt@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/02/15 20:06:59 by rlutt             #+#    #+#             */
-/*   Updated: 2017/03/09 12:49:59 by rlutt            ###   ########.fr       */
+/*   Updated: 2017/04/05 10:45:33 by rlutt            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,8 +23,8 @@ void				ft_putchar(char c);
 void				ft_putstr(char *str);
 void				ft_putstrrev(char *str);
 void				ft_putstr_fd(char const *s, int fd);
-void				ft_putnbr_fd(int n, int fd);
-void				ft_putnbr(int n);
+void				ft_putnbr_fd(intmax_t n, int fd);
+void				ft_putnbr(intmax_t n);
 void				ft_putendl_fd(char const *s, int fd);
 void				ft_putendl(char const *s);
 void				ft_putchar_fd(char c, int fd);
